+++ lib/src/widgets/bs_data_cell.dart
@@ -4,14 +4,21 @@ class BsDataCell extends StatelessWidget {
   const BsDataCell(
     this.child, {
     Key? key,
+    this.padding = const EdgeInsets.all(15.0),
+    this.alignment,
   }) : super(key: key);
 
   final Widget child;
 
+  final EdgeInsets padding;
+
+  final AlignmentGeometry? alignment;
+
   @override
   Widget build(BuildContext context) {
     return Container(
-      padding: EdgeInsets.all(15.0),
+      padding: padding,
+      alignment: alignment,
       child: Row(
         children: [
           Expanded(child: Container(child: child))